adding logic to Profile.js
we are using the withAuthenticator HOC to render the authentication flow by wrapping the Profile component in the default export. This will show the sign-up/sign-in form if the user is not signed in, and if the user is signed in will show the UI with the user’s profile details.